Add docs + rename ConfigNullProvider to EmptySecretProvider [ci fast]
