--- modules/nextflow/src/main/groovy/nextflow/secret/ConfigNullProvider.groovy
@@ -1,42 +1,0 @@
-/*
- * Copyright 2013-2024, Seqera Labs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package nextflow.secret
-
-import groovy.transform.CompileStatic
-
-/**
- * Specialization of the null secrets provider that is used to
- * determine whether secrets are required in the config.
- *
- * @author Ben Sherman <bentshermann@gmail.com>
- */
-@CompileStatic
-class ConfigNullProvider extends NullProvider {
-
-    private boolean accessed
-
-    @Override
-    Secret getSecret(String name) {
-        accessed = true
-        return new SecretImpl(name, '')
-    }
-
-    boolean usedSecrets() {
-        return accessed
-    }
-}
